+++ RELEASE_NOTES.md
@@ -1,5 +1,8 @@
 ### New in 0.69 (not released yet)
 
+* New: Support for async read model updates (`IAmAsyncReadModelFor`).
+  You can mix and match asynchronous and synchronous updates, 
+  as long as you don't subscribe to the same event in both ways.
 * Fix: Added the schema `dbo` to the `eventdatamodel_list_type` in script `0002 - Create eventdatamodel_list_type.sql` for `EventFlow.MsSql`.
 
 ### New in 0.68.3728 (released 2018-12-03)